+++ eng/common.sh
@@ -8,37 +8,10 @@ require_command() {
   }
 }
 
-
 semantic_type_model_package_ids() {
-  cat <<'PACKAGES'
-SemanticTypeModel.Abstractions
-SemanticTypeModel.Core
-SemanticTypeModel.JsonSchema
-SemanticTypeModel.DotNet
-SemanticTypeModel.Generators
-SemanticTypeModel.DependencyInjection
-SemanticTypeModel.PowerBI
-SemanticTypeModel.EFCore
-SemanticTypeModel.EFCore.Generators
-SemanticTypeModel.SystemTextJson
-SemanticTypeModel.Configuration
-SemanticTypeModel.Configuration.Generators
-PACKAGES
+  dotnet run --project eng/Engineering.Commands/Engineering.Commands.csproj -- package-ids
 }
 
 semantic_type_model_package_projects() {
-  cat <<'PROJECTS'
-src/SemanticTypeModel.Abstractions/SemanticTypeModel.Abstractions.csproj
-src/SemanticTypeModel.Core/SemanticTypeModel.Core.csproj
-src/SemanticTypeModel.JsonSchema/SemanticTypeModel.JsonSchema.csproj
-src/SemanticTypeModel.DotNet/SemanticTypeModel.DotNet.csproj
-src/SemanticTypeModel.Generators/SemanticTypeModel.Generators.csproj
-src/SemanticTypeModel.DependencyInjection/SemanticTypeModel.DependencyInjection.csproj
-src/SemanticTypeModel.PowerBI/SemanticTypeModel.PowerBI.csproj
-src/SemanticTypeModel.EFCore/SemanticTypeModel.EFCore.csproj
-src/SemanticTypeModel.EFCore.Generators/SemanticTypeModel.EFCore.Generators.csproj
-src/SemanticTypeModel.SystemTextJson/SemanticTypeModel.SystemTextJson.csproj
-src/SemanticTypeModel.Configuration/SemanticTypeModel.Configuration.csproj
-src/SemanticTypeModel.Configuration.Generators/SemanticTypeModel.Configuration.Generators.csproj
-PROJECTS
+  dotnet run --project eng/Engineering.Commands/Engineering.Commands.csproj -- package-projects
 }